refactor: moved commands to dedicated file
